refactor(webrtc): deduplicate _varint.py — use libp2p.utils.varint
Deletes libp2p/transport/webrtc/_varint.py and migrates stream.py and
signaling.py to import from the shared libp2p.utils.varint utilities:

- encode_uvarint (identical algorithm, both files)
- decode_varint_with_size (same (value, consumed) tuple, replaces
  decode_uvarint in stream.py)
- decode_uvarint_from_stream (byte-by-byte async reader, replaces the
  private _read_uvarint in signaling.py; error wrapping in
  read_signaling_message preserves WebRTCSignalingError at call sites)

Updates three test files to import the canonical utilities directly.
No wire-format or protocol change.

Closes #1355.
