+++ tests/core/transport/webrtc/test_connection.py
@@ -124,7 +124,7 @@ class TestMessageRouting:
     async def test_on_channel_message_routes_to_stream(self):
         conn = _make_connection()
         stream = await conn.open_stream()
-        from libp2p.transport.webrtc._varint import encode_uvarint
+        from libp2p.utils.varint import encode_uvarint
         from libp2p.transport.webrtc.pb.webrtc_pb2 import Message
 
         proto = Message(message=b"test-data").SerializeToString()
