+++ tests/core/transport/webrtc/test_signaling.py
@@ -13,10 +13,10 @@ from __future__ import annotations
 import pytest
 import trio
 
-from libp2p.transport.webrtc._varint import encode_uvarint as _encode_uvarint
+from libp2p.utils.varint import decode_uvarint_from_stream as _read_uvarint
+from libp2p.utils.varint import encode_uvarint as _encode_uvarint
 from libp2p.transport.webrtc.signaling import (
     SignalingSession,
-    _read_uvarint,
     read_signaling_message,
     write_signaling_message,
 )
